data_collection/requirements.txt
beautifulsoup4==4.9.1
cloudant==2.13.0
JPype1==1.0.1
numpy==1.19.1
opencv-python==4.3.0.36
pdf2image==1.13.1
pysolr==3.9.0
pytesseract==0.3.4
simplejson==3.17.0
tesseract==0.1.3
tqdm==4.48.0